add register validate

+++ handlers/excel.py
@@ -29,22 +29,29 @@ async def get_excel(message: Union[types.Message, types.CallbackQuery],
         df_sql = await conn.run_sync(read_sql_query, stmt)
         df_sql = df_sql.astype({'date': str})
 
-    filename = str(datetime.now()) + '.xlsx'
-    df_sql.to_excel(filename)
-
-    try:
-        await message.answer_document(FSInputFile(filename))
-        await message.delete()
+    if df_sql.empty:
+        if isinstance(message, types.Message):
+            await message.answer('В базе пока нет данных о состоянии сотрудников')
+        else:
+            await message.answer('В базе пока нет данных о состоянии выбранного сотрудника')
     
-    except AttributeError:
-        message: types.CallbackQuery
-        await message.message.answer_document(FSInputFile(filename))
-        await message.message.delete()
+    else:
+        filename = str(datetime.now()) + '.xlsx'
+        df_sql.to_excel(filename)
+        try:
+            message: types.Message
+            await message.answer_document(FSInputFile(filename))
+            await message.delete()
+        
+        except AttributeError:
+            message: types.CallbackQuery
+            await message.message.answer_document(FSInputFile(filename))
+            await message.message.delete()
 
-    if os.path.isfile(filename):
-        os.remove(filename)
+        if os.path.isfile(filename):
+            os.remove(filename)
 
-    # print(df_sql['date'])
+        # print(df_sql['date'])
 
 
 @excel_router.message()
